+++ tests/WorkerTest.php
@@ -6,8 +6,12 @@ namespace AvtoDev\AmqpRabbitLaravelQueue\Tests;
 
 use Mockery as m;
 use Illuminate\Support\Str;
+use Interop\Amqp\AmqpQueue;
+use Enqueue\AmqpExt\AmqpContext;
+use Enqueue\AmqpExt\AmqpConsumer;
 use Illuminate\Queue\QueueManager;
 use Illuminate\Queue\WorkerOptions;
+use Interop\Queue\SubscriptionConsumer;
 use AvtoDev\AmqpRabbitLaravelQueue\Queue;
 use Illuminate\Queue\Events\JobProcessed;
 use AvtoDev\AmqpRabbitLaravelQueue\Worker;
@@ -175,5 +179,62 @@ class WorkerTest extends AbstractTestCase
         $this->worker->daemon($this->queue_connection_name, 'default', new WorkerOptions(0, 32, -1, 3, 2));
         \usleep(8500);
         $this->assertSame(0, $queue->size()); // There is no 'jobs.failer', so, failed job should be 'deleted
+    }
+
+    /**
+     * @return void
+     */
+    public function testConsumerTagPrefix(): void
+    {
+        $manager = m::mock(QueueManager::class)
+            ->makePartial()
+            ->shouldReceive('connection')
+            ->andReturn($queue = m::mock(Queue::class))
+            ->getMock();
+        $queue
+            ->shouldReceive('getRabbitConnection')
+            ->andReturn($context = m::mock(AmqpContext::class))
+            ->getMock()
+            ->shouldReceive('getRabbitQueue')
+            ->andReturn(m::mock(AmqpQueue::class))
+            ->getMock()
+            ->shouldReceive('shouldResume')
+            ->andReturnFalse()
+            ->getMock();
+        $context
+            ->shouldReceive('createConsumer')
+            ->andReturn($consumer = m::mock(AmqpConsumer::class))
+            ->getMock()
+            ->shouldReceive('createSubscriptionConsumer')
+            ->andReturn($subscriber = m::mock(SubscriptionConsumer::class))
+            ->getMock()
+            ->shouldReceive('close')
+            ->getMock();
+
+        $prefix = Str::random(5);
+        $consumer
+            ->expects('setConsumerTag')
+            ->withArgs(function ($arg) use ($prefix) {
+                $this->assertMatchesRegularExpression("/^$prefix.*/", $arg);
+
+                return true;
+            })
+            ->getMock();
+        $subscriber
+            ->shouldReceive('subscribe')
+            ->getMock()
+            ->shouldReceive('consume')
+            ->getMock();
+
+        $worker = new Worker(
+            $manager,
+            $this->app->make(EventsDispatcher::class),
+            $this->app->make(ExceptionHandler::class),
+            static function () {
+            },
+            $prefix
+        );
+
+        $worker->daemon(Str::random(), 'default', new WorkerOptions());
     }
 }